hf-space-backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-text-splitters
llama-cpp-python
transformers
polygon-api-client 